Fixed: - calculate main stat equipment - create additional stat in rewards/purchases

+++ Assets/Project/Scripts/UI/Inventory/PurchaseRewardItemsView.cs
@@ -72,7 +72,7 @@ namespace UndergroundFortress.UI.Information
                 return;
             }
             
-            _shoppingService.Pay(_purchaseStaticData.moneyType, (int) _purchaseStaticData.price);
+            _shoppingService.Pay(_purchaseStaticData.moneyType, _purchaseStaticData.price);
             
             base.ClaimRewards();
         }
